requirements.txt
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
Flask==2.2.2
Flask-PyMongo==2.3.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.6.0
pymongo==4.3.3
Werkzeug==2.2.2
